refactor: change structure and add cors
