+++ backend/package.json
@@ -25,6 +25,7 @@
     "license": "MIT",
     "devDependencies": {
         "@types/bcrypt": "^5.0.0",
+        "@types/cors": "^2.8.12",
         "@types/express": "^4.17.13",
         "@types/jest": "^27.0.3",
         "@types/jsonwebtoken": "^8.5.6",
@@ -52,6 +53,7 @@
         "@babel/preset-typescript": "^7.16.5",
         "bcrypt": "^5.0.1",
         "body-parser": "^1.19.1",
+        "cors": "^2.8.5",
         "dotenv": "^10.0.0",
         "express": "^4.17.1",
         "jest": "^27.4.5",
