+++ backend/src/db/providers/authProvider.ts
@@ -1,5 +1,5 @@
 import UserModel from 'db/models/User';
-import { IUser } from 'interfaces/db/entities';
+import { IUser } from 'interfaces/entities/user';
 
 const authProvider = async (login: string) => {
   const dbUser = await UserModel.findOne({ username: login }).lean();
