+++ backend/src/interfaces/entities/user.d.ts
