+++ backend/src/routing/routes.ts
@@ -1,0 +1,13 @@
+import { Router } from 'express';
+
+import { Routes } from 'enums/routesEnum';
+
+import accountRouter from './accountRouter';
+import userRouter from './userRouter';
+
+const routers = Router();
+
+routers.use(`${Routes.account}`, accountRouter);
+routers.use(`${Routes.users}`, userRouter);
+
+export default routers;
