+++ backend/src/utils/dto/dtoUtils.ts
@@ -1,5 +1,5 @@
 import { ITokens } from 'interfaces/auth/authInterfaces';
-import { IUser } from 'interfaces/db/entities';
+import { IUser } from 'interfaces/entities/user';
 
 const generateInitialDto = (user: IUser, tokens?: ITokens) => {
   const { _id: id, passwordHash, refreshToken, ...userDataToClient } = user;
